# Improve performance of unknown_interval