``` r
library(tsibble)
itvl <- interval(tourism)

unknown_interval_new <- function(x) {
  for(y in x) if(y) return(FALSE)
  !is_empty(x)
}

unknown_interval <- tsibble:::unknown_interval

bench::mark(
  unknown_interval(itvl),
  unknown_interval_new(itvl)
)
#> # A tibble: 2 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 unknown_interval(itvl)       3.59µs   3.98µs   216170.        0B     21.6
#> 2 unknown_interval_new(itvl) 466.01ns 493.95ns  1903858.        0B      0
```

<sup>Created on 2019-08-23 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>